re-exporting settings components
