+++ src/app/chat/page.tsx
@@ -18,6 +18,8 @@ import {
   type MessageFormat,
 } from "@/misc/types";
 
+import { IModal } from "@/misc/types/modals";
+
 import Profile from "@/components/user/profile";
 import ChatMenu from "@/components/chat/menu/chatMenu";
 import Conversation from "@/components/chat/conversation";
@@ -33,20 +35,20 @@ import ChatLists from "@/components/chat/chatLists";
 import Search from "@/components/common/search";
 import ChatTopBar from "@/components/chat/chatTopBar";
 import ContactInfo from "@/components/contactInfo/contact";
-import GeneralLayout from "@/components/common/generalLayout";
-import Settings from "@/components/settings";
+import Settings, {
+  ThemeSwitch,
+  Security,
+  KeyboardShortcuts,
+  Help,
+  RequestInfo,
+  Privacy,
+  SettingNotifications,
+} from "@/components/settings";
 import {
   IShowComponent,
   ShowComponentLabel,
 } from "@/misc/types/renderComponent";
-import KeyboardShortcuts from "@/components/settings/keyboardShortcuts";
-import { IModal } from "@/misc/types/modals";
-import ThemeSwitch from "@/components/settings/theme";
-import SettingNotifications from "@/components/settings/notifications";
-import Privacy from "@/components/settings/privacy";
-import Security from "@/components/settings/security";
-import Help from "@/components/settings/help";
-import RequestInfo from "@/components/settings/requestInfo";
+
 
 registerServiceWorker();
